backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.29
asyncpg>=0.29.0
pydantic>=1.10.14
python-jose>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
email-validator>=2.1.1
python-multipart>=0.0.9
